+++ APIPRODUCT/Patterns/StrategyPatterns/TaxOfficeManager/MediumTaxStrategy.cs
@@ -9,7 +9,7 @@ namespace Patterns.StrategyPatterns.TaxOfficeManager
 {
     public class MediumTaxStrategy : ITaxStrategy
     {
-        public List<int> CalculateTaxs(List<Employee> employees)
+        public List<double> CalculateTaxs(List<Employee> employees)
         {
             throw new NotImplementedException();
         }
